Return ENOMEM, not -ENOMEM
This is userland, not kernel-land.

Technically this changes the API, which I normally would be loathe
to do, but I don't know that I have ever actually seen malloc() fail,
set errno to ENOMEM, much less the program getting this behavior to
actually do something sensible with it, so in this case I am not
really too worried about it.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>