Allow choosing between float or double at compile time
By default, it will use double, but if you add -DOSNFLOAT=float
to CFLAGS, it will use floats.  This could in theory be faster
but I don't actually know if it is, and there could be other problems.

At a glance, it appears to work with floats, by which I mean only that
the three test images seem alright.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>